overlay LMO to TRTL src dir

+++ src/Common/IInputStream.h
@@ -17,7 +17,6 @@
 
 #pragma once
 
-#include <cstdint>
 #include <cstddef>
 
 namespace Common {
@@ -25,7 +24,7 @@ namespace Common {
 class IInputStream {
 public:
   virtual ~IInputStream() { }
-  virtual uint64_t readSome(void* data, uint64_t size) = 0;
+  virtual size_t readSome(void* data, size_t size) = 0;
 };
 
 }
