+++ src/Common/StdInputStream.h
@@ -26,7 +26,7 @@ class StdInputStream : public IInputStream {
 public:
   StdInputStream(std::istream& in);
   StdInputStream& operator=(const StdInputStream&) = delete;
-  uint64_t readSome(void* data, uint64_t size) override;
+  size_t readSome(void* data, size_t size) override;
 
 private:
   std::istream& in;
