+++ src/Common/StdOutputStream.h
@@ -26,7 +26,7 @@ class StdOutputStream : public IOutputStream {
 public:
   StdOutputStream(std::ostream& out);
   StdOutputStream& operator=(const StdOutputStream&) = delete;
-  uint64_t writeSome(const void* data, uint64_t size) override;
+  size_t writeSome(const void* data, size_t size) override;
 
 private:
   std::ostream& out;
